Improve getpass() usage
As can be seen from: http://unixhelp.ed.ac.uk/CGI/man-cgi?getpass+3
getpass() is a deprecated function and it's advised not to use
it.

What's more all signals are disabled so (SIGINT, SIGQUIT, SIGSTOP,
SIGTSTOP) will do nothing and as such if the user wants to quit the
program while typing the password in the password loop he is out of
luck.

With this commit we get a manual version of getpass(), inspired by this
SO answer: http://stackoverflow.com/a/1196696/110395