Flush ethash error before thread terminate
If there is an error in the ethash functions, the worker thread that
called them will call terminate due to an unhandled exception. Now we
make sure that at least something appears in the log before that.

If in the future we can flush our LogOutputstream then we should use
that instead.