Potential fix for EthashAux race condition
Some EthashAux functions are protected by mutexes. These mutexes are
normal members of the EthashAux class which itself is a singleton. It
can happen that multiple threads will get inside the singleton creation
`get()` functions when requesting a lock, but the lock will not be
acquired due to the singleton not yet existing.

This leads to multiple threads going in the same code and causing double
free corruption.

Making the mutexes static members of the singleton should hopefully
address the issue