Update m_working state in Client for private chain mining
When mining on a private chain the mining would get stuck and not
start. The reason was that the difficulty was zero. If we move the
mining state from m_premine to m_working the difficulty will get updated
properly and we will be able to mine in a private chain.