Fix deadlock in the mining code
A deadlock was introduced in the mining code. The `x_minerWork` lock was
being locked and then to release said lock, it was waiting for
stopWorking().

Afterwards various threads were also waiting on the `x_minerWork` to be
released and we went into a deadlock. Calling all miners before start with setWork()
of an empty WorkPackage will ensure that stopWorking() is called.