Cmake OpenCL discovery should now work properly
Using find_package() inside a cmake function will unfortunately set the
`XXX_FOUND` variable only for the local function scope. So for custom
find/use functions we need to make sure that we propagate the FOUND
variable if we rely on it as we do with the `OpenCL_FOUND` one.