Add an option to disable using miniupnpc if found
As people have reported, later versions of the miniupnpc library seem to
be incompatibe with the way we use it and that leads to compile errors.

We need to:

1. properly handle a version check in cmake
2. Maybe abstract the interface to the library to take into account
different library versions?

Until then the user should now be able to add `-DNO_MINIUPNPC=1` to the
cmake configure command to disable it even if it's found.