add removed unique_ptr.get() back
The get() is needed here because llvm::MemoryBuffer::getFile() does not
actually return a simple unique_ptr but instead returns an `ErrorOr`
construct and as such the compiler needs to first get the pointer and
then get the function of the file. Else it does not compile.

Here is the error I was getting
```
[ 13%] Building CXX object evmjit/libevmjit/CMakeFiles/evmjit.dir/Endianness.cpp.o
/home/lefteris/ew2/libethereum/evmjit/libevmjit/Cache.cpp:121:58: error: no member named 'getBuffer' in 'std::unique_ptr<llvm::MemoryBuffer, std::default_delete<llvm::MemoryBuffer> >'
                g_lastObject = llvm::MemoryBuffer::getMemBufferCopy(r->getBuffer());
```

Here is the function signature
```
  static ErrorOr<std::unique_ptr<MemoryBuffer>>
  getFile(const Twine &Filename, int64_t FileSize = -1,
          bool RequiresNullTerminator = true, bool IsVolatileSize =
		  false);
```