Patch for crypto++ 5.6.2-4 in ArchLinux
As described in https://github.com/ethereum/libethereum/issues/141 this
is an incompatibility introduced by the latest package in ArchLinux. If
using gcc this patch will fix the issue and compile will continue
smoothly.

It may break compilation for other platforms and as such please test
before merging. If it does we need to add some platform detection code
in the cmake file.