Fix missed framework contents
Due to a known bug in QT5, we have to fix up the linked
libraries if we're going to have a (distributable)
result of building. This is already done to a great
extent, but a .app that is included in one of the
frameworks is missed by the current fix. This commit
catches it, and (hopefully) any others that slip through
in the future.

See also:
https://github.com/ethereum/homebrew-ethereum/issues/46
https://github.com/Homebrew/homebrew/issues/47848

This removes the need to follow the workaround described
at

https://github.com/ethereum/webthree-umbrella/wiki/QTBUG-50155-workaround

and restores (I believe) `brew install cpp-ethereum` as a
command that will execute successfully on a clean machine.