Weakened off what "static linking" means when building for OS X. Apple explicitly does not support statically linked binaries (i.e., for the executable itself) on OS X. See https://developer.apple.com/library/mac/qa/qa1118/_index.html for more info. But it does support static linkage of the external libraries being used by that executable, which is almost as good. This is part of https://github.com/ethereum/webthree-umbrella/issues/495, where we are adding support for static builds to the codebase.
For OS X we currently only have partially-static library linkage too.   We would need to build the following from source to have .a files to link against instead of their dylibs which we get from Homebrew:

- jsoncpp
- json-rpc-cpp
- leveldb
- miniupnp
- gmp

Two further libraries (curl and zlib) ship as dylibs with the platform but again we could build from source and statically link these too.

RelWithDebugInfo for a statically linked OS X binary is 13MB.