Removed accidental unconditional dependency on CURL.
It appears that the nesting was always wrong, but we used to be able to get away with it when the code was using explicit find_package(), target_include_directories() and target_link_libraries() calls.   Now that the dependency onto CURL is a "normal" eth_use, the issue was exposed, resulting in an unwanted CURL dependency on any system which had CURL installed.