Temporarily forcing OS X binaries to Debug as a workaround for the Heisenbug.
Ideally we would only have built Mix in Debug, but it turns out that is close
to impossible because of the existing umbrella setup, so we'll do a
quick-and-dirty for the time being.   We will be able to do this better when
the repositories are restructured.

Fixes https://github.com/ethereum/webthree-umbrella/issues/546.