ethvm: Sort out input data and code arguments
The code can be provided by --code argument or by a binary file or by the standard input if the file name is "-".

The input data can be provided by --input arugment followed by hex encoded data.

If input data is provided but no code, a contract is created out of the data.