Suppressing deprecation warnings for SHA3_256 and SHA3_512 classes in CryptoPP. These attributes were added after the 5.6.3 release.
From that header file ...

"The Crypto++ SHA-3 implementation dates back to January 2013 when NIST
selected Keccak as SHA-3. In August 2015 NIST finalized SHA-3, and it
was a modified version of the Keccak selection. Crypto++ 5.6.2 through
5.6.4 provides the pre-FIPS 202 version of SHA-3; while Crypto++ 5.7
and above provides the FIPS 202 version of SHA-3.

See also http://en.wikipedia.org/wiki/SHA-3"

This means that we will never be able to move to the CryptoPP-5.7.x
series of releases, because Ethereum requires Keccak, not the final
SHA-3 standard algorithm.   We are planning to migrate cpp-ethereum
off CryptoPP anyway, so this is unlikely to be a long-standing issue.

See https://github.com/ethereum/cpp-ethereum/issues/3088.