Disable various EIP-8 tests for Windows.
They appear to be consistently broken on Windows, so we will
disable them for the time being, so that we have "clean" tests to iterate on.
With any tests in a broken state, the automation is not useful, because
we do not see the working -> broken transition occur.

See https://github.com/ethereum/cpp-ethereum/issues/3273