Restore OS X Yosemite automation.
In August 2016, the 'carthage' package added a requirement for a full Xcode 7.3 install, not just command-line build tools, because it needs Swift 2.2. This requirement is a complete non-starter for TravisCI, where we have no ability or desire to do that installation. We aren't using 'carthage' ourselves, so we just pin it here prior to 'brew update'.

https://github.com/Homebrew/homebrew-core/issues/3996