Document running container as user (non-root) #3053
There is no stable way to define a user of the host-system inside the
Dockerfile. Therefore I describe how to *run* the container as the current
user. I also added a convenience script to `scripts/`, that users can install
in their PATH, that will behave just like a local installation.

The other option (that I did not follow here) would be `--userns-remap`, but
this needs more setup and is only supported in newer docker installations.