Update generating_tests.rst
@pirapira - proposing some minor formatting/grammatical/spelling improvements to the test documentation. Also added a link to a `cpp-ethereum` files for easier navigation.

I noticed that the `.rst` does not convert to markdown easily.
As an example of a potential source of confusion:
![](https://puu.sh/wncYN/a7eb6e0e53.png)
Sometimes code snippets are not shown, because github-flavored markdown requires different syntax.

Is there any interest in doing a full markdown conversion for this document, since it's being linked to in @souptacular 's [Metropolis Testing Guide](https://gist.github.com/Souptacular/fd197b1fac7c6d2660b0bef27a33ed40)? (which is in markdown)

Let me know, I'm happy to help with this.