Fixes graceful termination when `aleth --test` is invoked but required a `SIGKILL` to terminate.
This was because the UnixSocketServer requires at least one accepting
connection *after* SIGTERM/SIGINT has been received, which usually never
is the case.

This PR implements a trivial workaround by first waiting for up to N
msecs (N=1000 for now) for any input to arrive, and if not, it'll just
retry, hence, evaluating m_running, resulting in `aleth --test` to
properly terminate upon receival of `SIGINT` or `SIGTERM`.