Ensure unit tests only use MemoryDB
Move the setDatabaseKind(DatabaseKind::MemoryDB) call from TestOutputHelper::initTest to the TestOutputHelperFixture constructor. The problem with the previous location was that initTest() is executed for both the consensus tests and the unit tests and we want to only force MemoryDB for the unit tests. This new location is more in line with the approach used in the consensus tests (the database kind is set in each of their fixture constructors e.g. TransactionTestFixture).