Fix bc insert test and make db setting global
Fix the BlockChainInsertTests/bcBasicInsert test failure by modifying BlockChain::dumpDatabase function to generate a string of sorted data rather than relying on the underlying DB implementation to order the data.

Make the DatabaseKind setting global rather than specific to the consensus tests and the unit tests - this results in a much simpler design.