Prevent a host from adding itself to its own node table (#5409)
Update NodeTable::addNode to check if the node being added to the node table is one's self and if so, skip adding the node to the node table.

Add unit test which adds host to node table. Unit test adds the host to the node table and checks the node table's node count both before and after adding the host and verifies that they're equal.