Fix failing peer unit tests
The failing tests create multiple node table instances which listen on the same port (port 0). The node table instances listen on the same port since we initialize them from the network config which the tests create with listen port == 0. Note that this doesn't affect the hosts since when the host binds its tcp acceptor to an endpoint (in Network::tcp4Listen()), if a port of 0 is detected the listen port will be set to a default port (30303) or if that's in use, a random port which is then returned to the host (host::startedWorking()). To fix the failing tests, we update the network config with this returned port (which we then pass to the node table ctor).