Add localhost peer test cases
Add 2 localhost peer test cases, both which verify that a node with a localhost ip cannot be added as a peer (when 'allowLocalDiscovery' is false).

These tests validate 2 different code paths - the first validates the Host::addNode() code path and the second validates the Host::requirePeer() code path.

Note that we don't need to verify the Host::addPeer() code path since that calls down to either Host::requirePeer() or Host::addNode() so it's already covered by the 2 new tests.