Don't involve nodes with localhost IPs in discovery process (#5429)
Prevent nodes with localhost IPs from being involved in the discovery process (unless "--allow-local-discovery" is set to true). Also add a couple of unit tests - one which verifies that a node with a localhost IP cannot be added to the node table when --allow-local-discovery is false (default behavior) and one which verifies that a peer cannot be made required when the network isn't running.