Streamline host shutdown logic
Prior to these changes, host shutdown would involve the app thread setting m_run to false and waiting on a condition variable until the timer used to schedule the host's main work loop (Host::run) was destroyed on the network thread. Note that as a part of shutting down the network thread, the node table was destroyed and the boost io service was stopped (in that order). Because the timer was accessed on two threads, it was protected by a mutex. Additionally, because the node table was destroyed before the io service and the node table scheduled discovery round execution via boost deadline timers which used the node table's captured this pointer, the node table could be destroyed while there was an outstanding scheduled execution which would cause a use-after-free.

I've simplified host shutdown by moving io service stoppage to the app thread (Host::stop) and letting the node table and timer be destroyed automatically when the Host instance is destroyed. Since the timer is now only ever accessed on the network thread, I've removed the associated mutex. Since the app thread no longer needs to wait while the network thread shuts down, we no longer need the timer condition variable either (m_timerReset).

This should also address the use-after-free since the io service is now stopped before node table is destroyed.