Call determinePublic() before node table ctor
We don't need to call determinePublic() before host::runAcceptor() because host::runAcceptor() only uses the public tcp endpoint data in a single log message, and this log message is redundant anyway since we already include public tcp endpoint info in the enode URL that we log soon after aleth launch. As such, the determinePublic() call can be moved from the if/else to right before the node table ctor is called for slightly less code duplication.