Fix capability use-after-free
A use-after-free can occur if capability handlers are still in the boost ioservice handler queue when the Host is destroyed, because capability handlers capture a shared ptr to the capability and destroying the capability results in Host::forEachPeer() being called which iterates over sessions which have already been destroyed (since the Host's session map  - m_sessions - is defined after the io service in Host.h).

I think the root of the issue is that the capability handlers capture a shared_ptr of the capability - this doesn't make sense since the host owns the capability lifetime and manages resources required by the capabilities (e.g. network sockets/sessions), so it doesn't make sense to use a shared_ptr in the handler to keep capabilities alive longer than the host. As such, I've changed the shared_ptr reference to a raw pointer reference. This is safe with no possibility of UAF because we check the capability's timer expiration time (and the timer is a captured shared_ptr so it's guaranteed to be alive) before calling into the capability's background work function and the capability timers are cancelled on host shutdown, so we will avoid calling into the capability when it has been destroyed.

Another issue is that the timer cancellations aren't processed on shutdown if there aren't any peers or incoming connections, because the capabilities are cancelled (via lambdas posted to the network thread) after the io service has stopped and Aleth only polls the io service to clear out peers / connections. To address this, I've added a ioservice post call after stopCapabilities() in Host::doneWorking() to ensure that the cancellations are processed.