Don't manually poll the ioservice after stopping capabilities
Manually polling the ioservice on shutdown is redundant. The motivation behind doing this was to ensure that the posted capability timer cancellations (see Host::stopCapabilities) were processed during host shutdown. However, it turns out that there's no need to do this because the capability timers only need to be cancelled if more capability work loops are being scheduled, but more capability work loops are only scheduled during shutdown if we are disconnecting pending handshakes and/or peers, each of which result in the io service being polled anyway.