Log block queue status in BlockChainSync
Log the block queue status when BlockChainSync cannot sync with a peer due to syncing being paused and when it cannot request blocks due to the block queue being full.

Also convert some BlockQueue constants to constexpr and wrap checking for a paused sync in a helper function (BlockChainSync::isSyncPaused) in the interest of consistency since pausing a sync is done via a helper function (BlockChainSync::pauseSync).