Set new handshake failure reason on disconnect
Set new HandshakeFailureReason::DisconnectRequested failure reason when a disconnect packet is received instead of a hello packet during the rlpx handshake. Additionally, decode the disconnect reason if present and log it to the rlpx channel.

Note that this change will prevent the host from garbage-collecting nodes with which the handshake fails due to a disconnect being received instead of hello since we don't treat this condition as critical (see Peer::isUseless) - this is desirable because we frequently encounter nodes which send us a disconnect instead of hello but with which we can later successfully complete the handshake and sync.

Note that receiving a disconnect instead of a hello isn't a violation of the protocol since the RLPX handshake spec states that a disconnect can be sent at any time.