Check if session is connected after obtaining shared_ptr
There are a few places where we retrieve a session (stored in Host::m_sessions) but we don't check if the session is connected before using the session ptr. We need to do this because sessions are kept alive by shared_ptrs (shared_from_this) passed to boost ASIO handlers, and there are conditions which can occur where a session shared_ptr is still valid but the session has been disconnected. For example, take the case of if we've requested a Session disconnect (Session::disconnect) but the handler which executes after the socket write occurs (see Session::write) is sitting in the boost ASIO handler queue.