Remove reprocess RPC method and associated type
The reprocess admin RPC method is the only function which depends on the
old Block::populateFromChain functionality where it replays each tx in
the block. There's no need for the function so we can remove it.

Additionally, check for StateDB corruption in Block::populateFromChain
before setting the state root so that we can log a user-friendly error
message.