Fix race condition which causes sync to get stuck
The problem is that the knownFull -> !knownFull check in
BlockQueue::drain() doesn't catch cases where the block queue is
modified on a verifier thread - these cases are when a bad block is
detected and the block and all of its children are removed from
m_verifying, or when a block (with bytes) is moved from m_unverified to
m_verifying (without bytes).

The fix is to always fire the onRoomAvailable handler (renamed to onBlocksDrained) at the end of
doneDrain, and in the handler to only resume syncing if it's paused and
the block queue isn't full.